% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_these.R
\name{attach_these}
\alias{attach_these}
\title{Attach (and if necessary, install) packages}
\usage{
attach_these(required_packages)
}
\arguments{
\item{required_packages}{A character vector listing all packages the user wants to attach.}
}
\value{
A character vector.
}
\description{
Attach (and if necessary, install) packages
}
\examples{
required_packages <- c("rlang", "devtools")
}
