% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mailing_list.R
\name{create_mailing_list}
\alias{create_mailing_list}
\title{Create Mailing List}
\usage{
create_mailing_list(
  path = "/home/file-server/09_WA/08_Div/01_Adressen/Adressen_Gemeinden_Parteien.xlsx",
  sheets = c(2, 3),
  columns = c("Email", "Email")
)
}
\arguments{
\item{path}{Path to an xlsx file that contains email-addresses in multiple sheets. The output file will be written into the same directory as "mailverteiler_YYYY-MM-DD.xlsx".}

\item{sheets}{Index numbers or names of the sheets that contain email address data.}

\item{columns}{Name of the columns in the sheets that contain email address data.}
}
\value{
Creates a file "mailverteiler_YYYY-MM-DD.xlsx"
}
\description{
Create Mailing List
}
\examples{
create_mailing_list <- function(path = "/home/file-server/09_WA/08_Div/01_Adressen/Adressen_Gemeinden_Parteien.xlsx", sheets = c(2, 3), columns = c("Email", "Email"))
}
